refactor: moved visitors out of sub-packages
